refactor: rename a folder
