+++ app/components/typography/AuthHeadingText.tsx
